Fixed gem name
Rails calls `Bundler.require` on initialization.

For every gem in Gemfile, `Bundler.require` will require that
gem, by attempting to require according to the gems name.

This means that typically, gems must have a file under `lib` that
corresponds to the name of the gem.

This gem was expecting users to require with `elastic/site-search`,
which did not match the gem name. This means gems had to be
explicitly required in Rails and also is a bit of an anti-pattern for
Ruby.

This fixes that by adding the correct file, so that this gem will be
auto-required by Rails and also available via
`require elastic-site-search`, which remaining backwards compatible with
`require elastic/site-search`.